Use integer delay

+++ src/main/java/com/robertsoultanaev/javasphinx/routing/RelayInformation.java
@@ -1,4 +1,4 @@
 package com.robertsoultanaev.javasphinx.routing;
 
-public record RelayInformation(MixNode node, byte delay) {
+public record RelayInformation(MixNode node, int delay) {
 }
